Added brute rolling mode, clean up game creation form, added pk switching in wallet and after a game ends and hide private key on game screen
