+++ packages/nextjs/components/dicedemo/WinnerAnnouncement.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const WinnerAnnouncement = () => {
+  return <div>WinnerAnnouncement</div>;
+};
+
+export default WinnerAnnouncement;
